requirements.txt
PyGetWindow~=0.0.9
keyboard~=0.13.5
selenium~=4.9.0
pytesseract~=0.3.10
Pillow~=9.5.0
pywinauto~=0.6.8